Move FidesGlobal type to lib/consent-types
